requirements.txt
slack_sdk==3.19.0
requests==2.31.0
schedule==1.1.0
python-dotenv==1.0.1
beautifulsoup4==4.13.3
flask==3.1.0
notion-client==2.0.0
google-generativeai==0.8.4